RildService: First implementation of HIDL service for legacy vendor rild.
This is java implementation of the RIL HIDL service interface using the
pre Android O legacy vendor rild binaries.

This is a ril bringup. Work is still in progress...

(The legacy rild daemon initially crashes because rild-debug socket
was not found. This is because of the Android O internal renaming
of rild-debug to rild_debug. We fix this with a batch for init:

    init: Fix rild vendor crash for missing rild-debug socket: rild-debug
           is renamed to rild_debug in Oreo

    Change-Id: I3566aa70e7ddb6607a7b87013ace2cd61922e8cd

So the init batch is also needed.
)

Change-Id: I4352ba5e261fbbfba07423fe3329e8ffb02bebfc